pids2.rs -> pid.rs replacing it. Removed bin n_joint_arm and is now in examples
