+++ configs/controls/n_joint_arm2_params.toml
@@ -1,23 +1,19 @@
 [n_joint_arm_params]
-link_lengths = [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
+link_lengths = [1.0, 1.0, 1.0, 1.0, 1.0]
 link_angles = [
   0.785,
   0.785,
   0.785,
   0.785,
   0.785,
-  0.785,
-  0.785,
-  0.785,
-  0.785
 ] # radian for now. TODO: change to degrees
 
 [pid_params]
-kp = [2.5, 2.5, 2.5, 2.5, 2.5, 2.5, 2.5, 1.5, 1.5]
-ki = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-kd = [0.4, 0.4, 0.4, 0.4, 0.4, 0.4, 0.4, 0.4, 0.4]
-error_total_lower_bound = [-0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0.1]
-error_total_upper_bound = [0.1, 0.1, 0.1, 0.1, 0.1, -0.1, 0.1, 0.1, 0.1]
+kp = [2.5, 2.5, 2.5, 2.5, 2.5]
+ki = [0.0, 0.0, 0.0, 0.0, 0.0]
+kd = [0.4, 0.4, 0.4, 0.4, 0.4]
+error_total_lower_bound = [-0.1, -0.1, -0.1, -0.1, -0.1]
+error_total_upper_bound = [0.1, 0.1, 0.1, 0.1, 0.1]
 
 # Animations
 [window_params]
@@ -32,7 +28,7 @@ margin = 10
 label_size = 30
 label_font = "sans-serif"
 label_font_size = 15
-x_range = [-5.0, 10.0]
+x_range = [-5.0, 5.0]
 y_range = [-5.0, 5.0]
 
 [animation_params]
