+++ src/controls/mod.rs
@@ -1,3 +1,2 @@
 pub mod path_tracking;
 pub mod pid;
-pub mod pid2;
